Add Data Mocking for Standard Data Task and Alamofire (#4)
This change adds data mocking for the standard URLSession.dataTask(with:completionHandler:) method of URLSession as well as data mocking for Alamofire related delegate callbacks. The change includes the addition of a launch argument where data mocking may be enabled at app launch as well as a switch to enable and disable data mocking inside of the Chucker View Controller.